backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
mysqlclient
djangorestframework-simplejwt
django-cors-headers
boto3
django-storages
python-dotenv
gunicorn
django-allauth
dj-rest-auth
requests
cryptography